requirements.txt
scipy==1.6.2
librosa==0.8.1
seaborn==0.11.1
tensorflow==2.4.1
albumentations==1.0.3
torch_lr_finder==0.2.1
matplotlib==3.3.4
torchvision==0.8.0a0
torch==1.9.0
opencv_python==4.5.3.56
fastai==2.4.1
SoundFile==0.10.3.post1
tqdm==4.62.2
numpy==1.19.2
pandas==1.2.4
Keras==2.4.3
joblib==1.0.1
ipython==7.28.0
Pillow==8.3.2
scikit_learn==1.0
